frontend: storybook: Fix startup failure caused by node story helpers
Storybook startup failed with "Cannot access 'KubeObject' before
initialization" because baseMocks imported node/storyHelper, which
loaded node.ts at runtime only to access NODE_POOL_LABEL_KEYS and
type definitions.

Extract NODE_POOL_LABEL_KEYS into a dedicated nodeConstants module
and update storyHelper to use type-only imports, avoiding an
unnecessary runtime dependency on the Node implementation.

Re-export NODE_POOL_LABEL_KEYS from node.ts to preserve plugin and
consumer compatibility.
